joi validation

+++ controllers/task.controller.ts
@@ -1,6 +1,7 @@
 import { Request, Response } from "express";
 import { AppDatasource } from "../models/datasource";
 import { TaskEntity } from "../models/task.entity";
+import { taskValidation } from "../utils/validation";
 
 const taskRepository = AppDatasource.getRepository(TaskEntity);
 
@@ -28,11 +29,18 @@ const getTask = async (req: Request, res: Response) => {
 const postTask = async (req: Request, res: Response) => {
   try {
     const task = new TaskEntity();
-    task.name = req.body.name;
-    task.description = req.body.description;
 
-    await taskRepository.save(task);
-    return res.status(201).json({ msg: `Added task successfully` });
+    // Validation
+    const { error, value } = await taskValidation.validateAsync(req.body);
+    if (error) {
+      return res.status(400).json({ msg: `Validation failed..` });
+    } else {
+      task.name = req.body.name;
+      task.description = req.body.description;
+
+      await taskRepository.save(task);
+      return res.status(201).json({ msg: `Added task successfully` });
+    }
   } catch (error) {
     console.log(error);
     return res.status(500).json({ msg: `Unable to add task..` });
@@ -41,16 +49,28 @@ const postTask = async (req: Request, res: Response) => {
 // Update Task
 const updateTask = async (req: Request, res: Response) => {
   try {
-    const task: any = await taskRepository.findOneBy({
-      id: Number(req.params.id),
-    });
+    // validation
+    const { error, value } = await taskValidation.validateAsync(req.body);
+    if (error) {
+      return res.status(400).json({ msg: `Validation failed..` });
+    } else {
+      const task: any = await taskRepository.findOneBy({
+        id: Number(req.params.id),
+      });
 
-    task.name = req.body.name;
-    task.description = req.body.description;
+      if (!task) {
+        return res
+          .status(400)
+          .json({ msg: `Unable to find task with id:${req.body.id}` });
+      }
 
-    await taskRepository.save(task);
-    console.log(`Updated task with id ${req.params.id}`);
-    return res.status(200).json(task);
+      task.name = req.body.name;
+      task.description = req.body.description;
+
+      await taskRepository.save(task);
+      console.log(`Updated task with id ${req.params.id}`);
+      return res.status(200).json(task);
+    }
   } catch (error) {
     console.log(error);
     return res.status(500).json({ msg: `Unable to update task..` });
@@ -64,6 +84,11 @@ const deleteTask = async (req: Request, res: Response) => {
       id: Number(req.params.id),
     });
 
+    if (!task) {
+      return res
+        .status(400)
+        .json({ msg: `Unable to find task with id:${req.body.id}` });
+    }
     await taskRepository.remove(task);
     console.log(`Deleted task with id: ${req.params.id}`);
     return res.status(200).json(task);
